backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
redis>=5.0.0
pydantic-settings>=2.2.0
pydantic[email]>=2.6.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1,<5.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
python-multipart>=0.0.9
qrcode[pil]>=7.4.2
